+++ Parallel_tree_algorithm/lib/bindings/utils.js
