+++ Parallel_tree_algorithm/rules/singleRule.rule
