+++ tests/test_cuckoo_hash_table.py
@@ -1,8 +1,9 @@
 import unittest
-import trie_tree_parser.python.TrieTree.portNumberTrieTree as portnumbertrie
-import trie_tree_parser.python.TrieTree.policyFactory as policyFactory
-import trie_tree_parser.python.hashTable.cuckooHashTable as cuckooHashTable
+import Parallel_tree_algorithm.python.TrieTree.PolicyTrieTree as portnumbertrie
+import Parallel_tree_algorithm.python.TrieTree.policyBuilder as policyBuilder
+import Parallel_tree_algorithm.python.hashTable.cuckooHashTable as cuckooHashTable
 import logging
+
 
 class TestCuckooHashTable(unittest.TestCase):
     def setUp(self):
@@ -11,7 +12,7 @@ class TestCuckooHashTable(unittest.TestCase):
         self.tree2 = portnumbertrie.PortNumberTrieTree()
         treeList = [self.tree0,self.tree1,self.tree2]
 
-        self.policyFactory = policyFactory.PolicyFactory(treeList)
+        self.policyFactory = policyBuilder.PolicyBuilder(treeList)
         self.policyFactory.setSeed(311415)
         
         self.hashTable = cuckooHashTable.CuckooHashTable()
